requirements.txt
# requirements.txt
torch>=2.0.0
diffusers>=0.18.0
transformers>=4.30.0
accelerate>=0.20.0
pillow>=9.5.0
python-dotenv>=1.0.0